Move Globe assets to /assets
