--- static/globeCSS.css
@@ -1,90 +1,0 @@
-ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-  background-color: #333;
-}
-
-li {
-  border-right: 1px solid #bbb;
-  float: left;
-}
-
-li a {
-  display: block;
-  color: white;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-}
-
-li a:hover {
-  background-color: #111;
-}
-
-.active {
-  background-color: #4CAF50;
-}
-
-html {
-  height: 100%;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  color: #ffffff;
-  font-family: sans-serif;
-  font-size: 13px;
-  line-height: 20px;
-  height: 100%;
-}
-
-#info {
-  font-size: 11px;
-  position: absolute;
-  bottom: 5px;
-  background-color: rgba(0, 0, 0, 0.8);
-  border-radius: 3px;
-  right: 10px;
-  padding: 10px;
-}
-
-a {
-  color: #aaa;
-  text-decoration: none;
-}
-
-a:hover {
-  text-decoration: underline;
-}
-
-.bull {
-  padding: 0px 5px;
-  color: #555;
-}
-
-#title {
-  position: absolute;
-  top: 20px;
-  width: 320px;
-  left: 20px;
-  background-color: rgba(0, 0, 0, 0.2);
-  font: 20px/20px Georgia;
-  padding: 15px;
-}
-
-#ce span {
-  display: none;
-}
-
-#ce {
-  width: 107px;
-  height: 55px;
-  display: block;
-  position: absolute;
-  bottom: 15px;
-  left: 20px;
-  background: url(static/ce.png);
-}
