--- static/indexCSS.css
@@ -1,64 +1,0 @@
-ul {
-  list-style-type: none;
-  width: 100%;
-  top: 0;
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-  background-color: #333;
-}
-
-li {
-  border-right: 1px solid #bbb;
-  float: left;
-}
-
-li a {
-  display: block;
-  color: white;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-}
-
-li a:hover {
-  background-color: #111;
-}
-
-.active {
-  background-color: #4CAF50;
-}
-
-body {
-  height: 100%;
-  background-image: url("space.jpg");
-}
-
-#center {
-  font-family: 'Contrail One', cursive;
-  text-align: center;
-  color: white;
-  font-size: 200%;
-}
-
-p {
-  font-family: 'Contrail One', cursive;
-  text-align: center;
-  color: white;
-  font-size: 200%;
-  text-transform: uppercase;
-  font-weight: bold;
-  line-height: 1.5;
-  margin-bottom: 10px;
-  letter-spacing: 0.06em;
-}
-
-#signature {
-  position: absolute;
-  bottom: 0;
-}
-
-a {
-  text-decoration: none;
-  color: #42bcf4;
-}
