--- static/meteorites.json
@@ -1,1 +1,0 @@
-[50.775, 6.08333, 0.21, 56.18333, 10.23333, 0.720, 54.21667, -113, 0.8, 16.83333, -99.9, 0.7, -33.16667, -64.95, 0.780, 32.1, 71.8, 0.4239, 44.83333, 95.16667, 0.710, 44.21667, 61.1667, 0.8333, -31.6, -65.2333, 0.866, -30.8667, -64.55, 0.68, 16.39806, -9.5702, 0.4]
