--- static/updateCSS.css
@@ -1,37 +1,0 @@
-ul {
-  list-style-type: none;
-  margin: 1px;
-  padding: 1px;
-  overflow: hidden;
-  background-color: #333;
-}
-
-li {
-  border-right: 1px solid #bbb;
-  float: left;
-}
-
-li a {
-  display: block;
-  color: white;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-}
-
-li a:hover {
-  background-color: #111;
-}
-
-.active {
-  background-color: #4CAF50;
-}
-
-#nasa {
-  top: 20px;
-}
-
-body {
-  height: 100%;
-  background-image: url("space.jpg");
-}
