--- templates/updates.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Twitter Updates</title>
-
-    <link rel="stylesheet" type="text/css" href="static/updateCSS.css">
-  </head>
-
-
-  <body>
-    <!-- Base for navigation bar -->
-  <ul>
-    <li><a href="/">About</a></li>
-    <li><a href="/globe">Globe</a><li>
-    <li><a href="/map">Map</a></li>
-    <li><a class="active" href="/updates">Updates</a></li>
-  </ul>
-
-  <div id="nasa">
-<a class="twitter-timeline" data-height="600" data-theme="dark" href="https://twitter.com/NASA">Tweets by NASA</a> <script async src="//platform.twitter.com/widgets.js" charset="utf-8"></script>
-</div>
-
-  <a class="twitter-timeline" data-height="600" data-theme="dark" href="https://twitter.com/AsteroidWatch">Tweets by AsteroidWatch</a> <script async src="//platform.twitter.com/widgets.js" charset="utf-8"></script>
-  </body>
-</html>
